Actions and Auth MIddleware

+++ src/components/delete-button.tsx
@@ -1,22 +1,27 @@
-'use client'
+'use client';
 
-import { deleteCliente } from "@/actions/actions"
+import { deleteCliente } from '@/actions/actions';
 
 type Props = {
-  id: number
-}
+  id: number;
+};
 
-const DeleteButton = ({id}: Props) => {
+const DeleteButton = ({ id }: Props) => {
+  const handleDelete = async () => {
+    if (confirm('¿Estás seguro de eliminar este cliente?')) {
+      const res = await deleteCliente(id);
+      console.log( res )
+      if (res) {
+        alert(res);
+      }
+    }
+  };
+
   return (
-    <button
-      onClick={async () => {
-        await deleteCliente(id)
-      }}
-      className="bg-red-600 rounded-md px-4 py-2"
-    >
+    <button onClick={handleDelete} className="bg-red-600 rounded-md px-4 py-2">
       Eliminar
     </button>
-  )
-}
+  );
+};
 
-export default DeleteButton
+export default DeleteButton;
